+++ src/main/java/backend/wal/wal/todaywal/domain/view/Home.java
@@ -1,8 +1,9 @@
-package backend.wal.home.domain;
+package backend.wal.wal.todaywal.domain.view;
 
+import backend.wal.wal.todaywal.domain.aggregate.TodayWal;
+import backend.wal.wal.todaywal.domain.aggregate.ShowStatus;
 import backend.wal.wal.common.domain.WalCategoryType;
 import backend.wal.wal.common.domain.WalTimeType;
-import backend.wal.wal.todaywal.domain.aggregate.ShowStatus;
 import com.google.common.base.Objects;
 import lombok.Getter;
 
@@ -18,7 +19,7 @@ public final class Home {
     private final ShowStatus showStatus;
     private OpenStatus openStatus;
 
-    public Home(final Long todayWalId, final WalTimeType timeType, final WalCategoryType categoryType,
+    private Home(final Long todayWalId, final WalTimeType timeType, final WalCategoryType categoryType,
                 final String message, final ShowStatus showStatus) {
         this.todayWalId = todayWalId;
         this.timeType = timeType;
@@ -26,6 +27,16 @@ public final class Home {
         this.message = message;
         this.showStatus = showStatus;
         this.openStatus = OpenStatus.UNABLE;
+    }
+
+    public static Home of(final TodayWal todayWal) {
+        return new Home(
+                todayWal.getId(),
+                todayWal.getTimeType(),
+                todayWal.getCategoryType(),
+                todayWal.getMessage(),
+                todayWal.getShowStatus()
+        );
     }
 
     public void setOpenStatusBy(LocalDateTime now) {
